fix type annotation

+++ src/definitions/abi/fat_pointer.rs
@@ -60,7 +60,7 @@ impl FatPointer {
         if of {
             exceptions.set(FatPointerValidationException::DEREF_BEYOND_HEAP_RANGE, true);
         }
-        let (_, limit_exceeded) = ((1 << 30) - 1).overflowing_sub(end_non_inclusive);
+        let (_, limit_exceeded) = (((1 << 30) - 1) as u32).overflowing_sub(end_non_inclusive);
         if limit_exceeded {
             exceptions.set(FatPointerValidationException::DEREF_BEYOND_HEAP_RANGE, true);
         }
